rename ZVProgressHUD -> ProgressHUD, and use singleton instead class method.
