+++ Cartfile.resolved
@@ -1,1 +1,1 @@
-github "zevwings/ZVActivityIndicatorView" "0.1.6"
+github "zevwings/ZVActivityIndicatorView" "0.1.7"
