+++ ZVProgressHUD/Deprecated.swift
@@ -8,7 +8,193 @@
 
 import Foundation
 
-public extension ZVProgressHUD {
+@available(*, deprecated, renamed: "ProgressHUD", message: "renamed ProgressHUD will remove in future")
+public typealias ZVProgressHUD = ProgressHUD
+
+public extension ProgressHUD {
+    
+    @available(*, deprecated, message: "this attribute will remove in future, use singleton instead.")
+    class var displayStyle: DisplayStyle {
+        get {
+            return shared.displayStyle
+        }
+        set {
+            shared.displayStyle = newValue
+        }
+    }
+    
+    @available(*, deprecated, message: "this attribute will remove in future, use singleton instead.")
+    class var maskType: MaskType {
+        get {
+            return shared.maskType
+        }
+        set {
+            shared.maskType = newValue
+        }
+    }
+    
+    @available(*, deprecated, message: "this attribute will remove in future, use singleton instead.")
+    class var maxSupportedWindowLevel: UIWindow.Level {
+        get {
+            return shared.maxSupportedWindowLevel
+        }
+        set {
+            shared.maxSupportedWindowLevel = newValue
+        }
+    }
+    
+    @available(*, deprecated, message: "this attribute will remove in future, use singleton instead.")
+    class var fadeInAnimationTimeInterval: TimeInterval {
+        get {
+            return shared.fadeInAnimationTimeInterval
+        }
+        set {
+            shared.fadeInAnimationTimeInterval = newValue
+        }
+    }
+    
+    @available(*, deprecated, message: "this attribute will remove in future, use singleton instead.")
+    class var fadeOutAnimationTImeInterval: TimeInterval {
+        get {
+            return shared.fadeOutAnimationTImeInterval
+        }
+        set {
+            shared.fadeOutAnimationTImeInterval = newValue
+        }
+    }
+    
+    @available(*, deprecated, message: "this attribute will remove in future, use singleton instead.")
+    class var minimumDismissTimeInterval: TimeInterval {
+        get {
+            return shared.minimumDismissTimeInterval
+        }
+        set {
+            shared.minimumDismissTimeInterval = newValue
+        }
+    }
+    
+    @available(*, deprecated, message: "this attribute will remove in future, use singleton instead.")
+    class var maximumDismissTimeInterval: TimeInterval {
+        get {
+            return shared.maximumDismissTimeInterval
+        }
+        set {
+            shared.maximumDismissTimeInterval = newValue
+        }
+    }
+    
+    @available(*, deprecated, message: "this attribute will remove in future, use singleton instead.")
+    class var cornerRadius: CGFloat {
+        get {
+            return shared.cornerRadius
+        }
+        set {
+            shared.cornerRadius = newValue
+        }
+    }
+    
+    @available(*, deprecated, message: "this attribute will remove in future, use singleton instead.")
+    class var offset: UIOffset {
+        get {
+            return shared.offset
+        }
+        set {
+            shared.offset = newValue
+        }
+    }
+    
+    @available(*, deprecated, message: "this attribute will remove in future, use singleton instead.")
+    class var font: UIFont {
+        get {
+            return shared.font
+        }
+        set {
+            shared.font = newValue
+        }
+    }
+    
+    @available(*, deprecated, message: "this attribute will remove in future, use singleton instead.")
+    class var strokeWith: CGFloat {
+        get {
+            return shared.strokeWith
+        }
+        set {
+            shared.strokeWith = newValue
+        }
+    }
+    
+    @available(*, deprecated, message: "this attribute will remove in future, use singleton instead.")
+    class var indicatorSize: CGSize {
+        get {
+            return shared.indicatorSize
+        }
+        set {
+            shared.indicatorSize = newValue
+        }
+    }
+    
+    @available(*, deprecated, message: "this attribute will remove in future, use singleton instead.")
+    class var animationType: IndicatorView.AnimationType {
+        get {
+            return shared.animationType
+        }
+        set {
+            shared.animationType = newValue
+        }
+    }
+    
+    @available(*, deprecated, message: "this attribute will remove in future, use singleton instead.")
+    class var contentInsets: UIEdgeInsets {
+        get {
+            return shared.contentInsets
+        }
+        set {
+            shared.contentInsets =  newValue
+        }
+    }
+    
+    @available(*, deprecated, message: "this attribute will remove in future, use singleton instead.")
+    class var titleEdgeInsets: UIEdgeInsets {
+        get {
+            return shared.titleEdgeInsets
+        }
+        set {
+            shared.titleEdgeInsets = newValue
+        }
+    }
+    
+    @available(*, deprecated, message: "this attribute will remove in future, use singleton instead.")
+    class var indicatorEdgeInsets: UIEdgeInsets {
+        get {
+            return shared.indicatorEdgeInsets
+        }
+        set {
+            shared.indicatorEdgeInsets = newValue
+        }
+    }
+    
+    @available(*, deprecated, message: "this attribute will remove in future, use singleton instead.")
+    class var logo: UIImage? {
+        get {
+            return shared.logo
+        }
+        set {
+            shared.logo = newValue
+        }
+    }
+
+    @available(*, deprecated, message: "this attribute will remove in future, use singleton instead.")
+    class var logoSize: CGSize {
+        get {
+            return shared.logoSize
+        }
+        set {
+            shared.logoSize = newValue
+        }
+    }
+}
+
+public extension ProgressHUD {
     
     /// show a toast
     ///
@@ -16,6 +202,7 @@ public extension ZVProgressHUD {
     ///   - text: toast content
     ///   - superview: super view, if superview is nil, show on main window
     ///   - delayTimeInterval: the view will show delay the `delayTimeInterval`
+    @available(*, deprecated, message: "this method will remove in future, use singleton instead.")
     class func showText(_ text: String,
                         in superview: UIView? = nil,
                         on position: Position = .bottom,
@@ -30,6 +217,7 @@ public extension ZVProgressHUD {
     ///   - title: the success message remind users what you want
     ///   - superview: super view, if superview is nil, show on main window
     ///   - delayTimeInterval: the view will show delay the `delayTimeInterval`
+    @available(*, deprecated, message: "this method will remove in future, use singleton instead.")
     class func showSuccess(with title: String = "",
                            in superview: UIView? = nil,
                            on position: Position = .center,
@@ -45,7 +233,7 @@ public extension ZVProgressHUD {
     ///   - title: the error message remind users what you want
     ///   - superview: super view, if superview is nil, show on main window
     ///   - delayTimeInterval: the view will show delay the `delayTimeInterval`
-    
+    @available(*, deprecated, message: "this method will remove in future, use singleton instead.")
     class func showError(with title: String = "",
                          in superview: UIView? = nil,
                          on position: Position = .center,
@@ -60,6 +248,7 @@ public extension ZVProgressHUD {
     ///   - title: the warning message remind users what you want
     ///   - superview: super view, if superview is nil, show on main window
     ///   - delayTimeInterval: the view will show delay the `delayTimeInterval`
+    @available(*, deprecated, message: "this method will remove in future, use singleton instead.")
     class func showWarning(with title: String = "",
                            in superview: UIView? = nil,
                            on position: Position = .center,
@@ -74,6 +263,7 @@ public extension ZVProgressHUD {
     ///   - title: the message remind users what you want
     ///   - superview: super view, if superview is nil, show on main window
     ///   - delayTimeInterval: the view will show delay the `delayTimeInterval`
+    @available(*, deprecated, message: "this method will remove in future, use singleton instead.")
     class func show(with title: String = "",
                     in superview: UIView? = nil,
                     on position: Position = .center,
@@ -89,6 +279,7 @@ public extension ZVProgressHUD {
     ///   - title: the message remind users what you want
     ///   - superview: super view, if superview is nil, show on main window
     ///   - delayTimeInterval: the view will show delay the `delayTimeInterval`
+    @available(*, deprecated, message: "this method will remove in future, use singleton instead.")
     class func showProgress(_ progress: Float,
                             title: String = "",
                             in superview: UIView? = nil,
@@ -106,6 +297,7 @@ public extension ZVProgressHUD {
     ///   - superview: super view, if superview is nil, show on main window
     ///   - dismissAtomically: if `true` the `HUD` will dissmiss atomically
     ///   - delayTimeInterval: the view will show delay the `delayTimeInterval`
+    @available(*, deprecated, message: "this method will remove in future, use singleton instead.")
     class func showImage(_ image: UIImage,
                          title: String = "",
                          in superview: UIView? = nil,
@@ -123,6 +315,7 @@ public extension ZVProgressHUD {
     ///   - title: the message remind users what you want
     ///   - superview: super view, if superview is nil, show on main window
     ///   - delayTimeInterval: the view will show delay the `delayTimeInterval`
+    @available(*, deprecated, message: "this method will remove in future, use singleton instead.")
     class func showAnimation(_ images: [UIImage],
                              duration: TimeInterval = 0.0,
                              title: String = "",
@@ -139,6 +332,7 @@ public extension ZVProgressHUD {
     ///   - displayType: ZVProgressHUD.DisplayType
     ///   - superview: super view, if superview is nil, show on main window
     ///   - delayTimeInterval: the view will show delay the `delayTimeInterval`
+    @available(*, deprecated, message: "this method will remove in future, use singleton instead.")
     class func show(with displayType: DisplayType,
                     in superview: UIView? = nil,
                     on position: Position = .center,
@@ -152,7 +346,8 @@ public extension ZVProgressHUD {
     /// - Parameters:
     ///   - delay: the view will dissmiss delay the `delayTimeInterval`
     ///   - completion: dismiss completion handler
-    class func dismiss(delay: TimeInterval = 0.0, completion: ZVProgressHUDCompletionHandler? = nil) {
+    @available(*, deprecated, message: "this method will remove in future, use singleton instead.")
+    class func dismiss(delay: TimeInterval = 0.0, completion: ProgressHUDCompletionHandler? = nil) {
         shared.dismiss(with: delay, completion: completion)
     }
 }
