+++ ZVProgressHUD/IndicatorView.swift
@@ -9,7 +9,7 @@
 import UIKit
 import ZVActivityIndicatorView
 
-public class ZVIndicatorView: UIView {
+public class IndicatorView: UIView {
 
     public enum IndicatorType {
         case none
@@ -66,7 +66,7 @@ public class ZVIndicatorView: UIView {
     private var imageIndicaotorView: UIImageView?
     private var nativeActivityIndicatorView: UIActivityIndicatorView?
     private var flatActivityIndicatorView: ZVActivityIndicatorView?
-    private var progressIndicatorView: ZVProgressView?
+    private var progressIndicatorView: ProgressView?
     
     convenience init() {
         self.init(frame: .zero)
@@ -85,7 +85,7 @@ public class ZVIndicatorView: UIView {
 
 // MARK: - Override
 
-extension ZVIndicatorView {
+extension IndicatorView {
     
     override public var tintColor: UIColor! {
         didSet {
@@ -110,7 +110,7 @@ extension ZVIndicatorView {
 
 // MARK: - Private Method
 
-private extension ZVIndicatorView {
+private extension IndicatorView {
     
     func configImageIndicatorView(_ value: Any, animationDuration: TimeInterval = 0.0) {
 
@@ -137,7 +137,7 @@ private extension ZVIndicatorView {
         
         if let resource = value as? String {
             
-            guard let path = Bundle(for: ZVProgressHUD.self).path(forResource: "Resource", ofType: "bundle") else { return }
+            guard let path = Bundle(for: ProgressHUD.self).path(forResource: "Resource", ofType: "bundle") else { return }
             let bundle = Bundle(path: path)
             guard let fileName = bundle?.path(forResource: resource, ofType: "png") else { return }
             let image = UIImage(contentsOfFile: fileName)?.withRenderingMode(.alwaysTemplate)
@@ -177,7 +177,7 @@ private extension ZVIndicatorView {
         nativeActivityIndicatorView?.removeFromSuperview()
 
         if progressIndicatorView == nil {
-            progressIndicatorView = ZVProgressView(frame: .zero)
+            progressIndicatorView = ProgressView(frame: .zero)
             progressIndicatorView?.strokeColor = tintColor
         }
         
@@ -242,7 +242,7 @@ private extension ZVIndicatorView {
 
 // MARK: - ZVIndicatorView.IndicatorType
 
-extension ZVIndicatorView.IndicatorType {
+extension IndicatorView.IndicatorType {
     
     var resource: String {
         switch self {
@@ -277,8 +277,10 @@ extension ZVIndicatorView.IndicatorType {
     
     var showLogo: Bool {
         switch self {
-        case .progress, .indicator:
+        case .progress:
             return true
+        case .indicator(let animationType):
+            return animationType == .flat
         default:
             return false
         }
